+++ plugins/destination/postgresql/main.go
@@ -2,6 +2,7 @@ package main
 
 import (
 	"github.com/cloudquery/cloudquery/plugins/destination/postgresql/client"
+	"github.com/cloudquery/plugin-sdk/plugins"
 	"github.com/cloudquery/plugin-sdk/serve"
 )
 
@@ -9,6 +10,11 @@ const (
 	sentryDSN = "https://19d1257d36854a51b17c06614e76dc2d@o1396617.ingest.sentry.io/4503896817336320"
 )
 
+var (
+	Version = "Development"
+)
+
 func main() {
-	serve.Destination(client.New(), serve.WithDestinationSentryDSN(sentryDSN))
+	p := plugins.NewDestinationPlugin("postgresql", Version, client.New)
+	serve.Destination(p, serve.WithDestinationSentryDSN(sentryDSN))
 }